Fulfil.Dispense/scripts/requirements.txt
# Warning bq 3.31 removes py 3.8 and 3.9 support
google-cloud-bigquery==3.30.0